nom del projecte: tenda musical
descripcició: s'integra la funcionalitat per realitzar el pagament amb
paypali tarjetes de credit o debit i s'integra la funcionalitat per
accedir a la pantalla del pagament

+++ tendamusical-services/src/main/java/com/lostriver/tendamusicalservices/service/CarritoService.java
@@ -17,7 +17,7 @@ public interface CarritoService {
 	CarritoAlbum guardarAlbumsCarrito(ArtistaAlbumDTO artistaAlbumDTO, Carrito carrito, int cantidadAlbumSelecccionada);
 
 	
-	float calcularTotal(Carrito carrito);
+	double calcularTotal(Carrito carrito);
 	
 	void eliminarAlbumCarrito(CarritoAlbum carritoAlbum);
